requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0

# NLP libraries
gensim>=4.3.0  # Word2Vec
transformers>=4.30.0  # BERT/DistilBERT
torch>=2.0.0  # PyTorch
tokenizers>=0.13.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.20.0

# Data download (Kaggle)
kaggle>=1.5.0

# Utilities
python-dotenv>=1.0.0
